ttcurate: What the Package Does

Helper functions and guidance for curating datasets for the 'TidyTuesday' weekly data project. 'TidyTuesday' is a project by the 'Data Science Learning Community' in which they post a weekly dataset in a public data repository (https://github.com/rfordatascience/tidytuesday) for people to analyze and visualize. This package includes tools to explore past submissions, analyze dataset sizes, and follow best practices for data preparation and documentation.

See also

Author

Maintainer: Jon Harmon jonthegeek@gmail.com (ORCID)